Removed old stocd data schemas and models, updated few models with bigdecimal

+++ src/main/java/com/emplyee/employeedemo/model/employee/Employees.java
@@ -3,6 +3,7 @@ package com.emplyee.employeedemo.model.employee;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -44,10 +45,10 @@ public class Employees {
   private Employees manager;
 
   @Column(name = "salary")
-  private double salary;
+  private BigDecimal salary;
 
   @Column(name = "commission_pct")
-  private double commission_pct;
+  private BigDecimal commission_pct;
 
   @OneToMany(mappedBy = "manager")
   private List<Employees> subordinates;
