+++ src/main/java/com/emplyee/employeedemo/model/employee/Jobs.java
@@ -3,6 +3,7 @@ package com.emplyee.employeedemo.model.employee;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -18,10 +19,10 @@ public class Jobs {
   private String title;
 
   @Column(name = "min_salary")
-  private double min_salary;
+  private BigDecimal min_salary;
 
   @Column(name = "max_salary")
-  private double max_salary;
+  private BigDecimal max_salary;
 
   @OneToMany(mappedBy = "jobs")
   private List<Employees> employees;
